requirements.txt
aiohttp==3.8.1
aiohttp-cors==0.7.0
aioredis==1.3.1
aiosignal==1.2.0
anyio==3.5.0
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
autopep8==1.6.0
Babel==2.9.1
backcall==0.2.0
bleach==4.1.0
blessed==1.19.1
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
colorful==0.5.4
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
docutils==0.18.1
entrypoints==0.4
filelock==3.6.0
fonttools==4.29.1
frozenlist==1.3.0
fsspec==2022.2.0
google-api-core==2.6.0
google-auth==2.6.0
googleapis-common-protos==1.55.0
gpustat==1.0.0b1
grpcio==1.44.0
hiredis==2.0.0
idna==3.3
importlib-metadata==4.11.1
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.9.1
ipython==7.31.1
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
json5==0.9.6
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyter-server==1.13.5
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
keyring==23.5.0
kiwisolver==1.3.2
lxml==4.8.0
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
modin==0.12.1
more-itertools==7.0.0
msgpack==1.0.3
multidict==6.0.2
nbclassic==0.3.5
nbclient==0.5.11
nbconvert==6.4.2
nbformat==5.1.3
nest-asyncio==1.5.4
nltk==3.7
notebook==6.4.8
numpy==1.21.5
nvidia-ml-py3==7.352.0
opencensus==0.8.0
opencensus-context==0.1.2
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pkginfo==1.8.2
pluggy==1.0.0
prometheus-client==0.13.1
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.9.0
ptyprocess==0.7.0
py==1.11.0
py-spy==0.3.11
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
pytest==7.0.1
pytest-runner==6.0.0
python-dateutil==2.8.2
python-docx==0.8.10
python-Levenshtein==0.12.2
pytz==2021.3
PyYAML==6.0
pyzmq==22.3.0
ray==1.10.0
readme-renderer==33.0
redis==4.1.4
regex==2022.1.18
requests==2.27.1
requests-toolbelt==0.9.1
rfc3986==2.0.0
rsa==4.8
Send2Trash==1.8.0
simplify-docx==0.1.1
six==1.16.0
smart-open==5.2.1
sniffio==1.2.0
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
thefuzz==0.19.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
twine==3.8.0
typing_extensions==4.1.1
urllib3==1.26.8
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.2.3
wincertstore==0.2
wordcloud==1.8.1
wrapt==1.13.3
XlsxWriter==3.0.3
yarl==1.7.2
zipp==3.7.0
